+++ Source/Rr_Descriptor.c
@@ -309,8 +309,8 @@ void Rr_WriteBufferDescriptor(
     uint32_t ArrayIndex,
     VkDescriptorType Type,
     VkBuffer Handle,
-    uint32_t Size,
-    uint32_t Offset)
+    uint64_t Size,
+    uint64_t Offset)
 {
     RR_RETURN_IF_NO_LAYOUT(State);
